+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Upcoming
+* Potion of recall now works on owned entities (dogs, etc)
+  * Please note that if the owner is teleported before the animal, the animal's effect will not end and will not get tpd
 * Supplementaries' Speaker Block now requires a durium ingot instead of emerald
 
 ## 1.9.13